requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
cssselect2==0.8.0
Django==5.2.9
django-background-tasks==1.2.8
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
freetype-py==2.5.1
html5lib==1.1
idna==3.11
lxml==6.0.2
mysqlclient==2.2.7
numpy==2.2.6
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.3.3
pdfkit==1.0.0
pillow==12.0.0
pycairo==1.29.0
pycparser==2.23
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pypdf==6.4.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.5
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.4
svglib==1.6.0
tinycss2==1.5.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.6.1
webencodings==0.5.1
xhtml2pdf==0.2.17
